Post-login API: track recordMethod
This API is not supposed to be used outside of `onContinuePostLogin`, so
before we were throwing an error. However, in practice there is no error
thrown and we do need to implement a solution for testing
`onContinuePostLogin`, so for now I've opted to replace the exception
with the implementation.

This does mean it's possible for someone to misuse `recordMethod`, but
it's also the case that you can do this in the wild.

A possible future improvement might be to expliclity raise an error in
an `onContinuePostLogin` context, when we add support for it.