Type hinting support
In order to get this to work with commonjs modules, we tried a few
options. The most reliable proved to be moving all the exports into a
single file, as the `exports` map exhibited stubborn behaviour when
trying to get type definitions and commonjs working well together.

As part of this change, we had to add a `package.json` with a relative
reference to the parent library within the `examples`.